fix(ci): python-coverage-shards — keep pr_monitor + test parts under the 1500-line guardrail
This PR pushed src/awf/runtime/pr_monitor.py (1478→1507) and
tests/unit/runtime/test_pr_monitor_parts/test_pr_monitor_part_003.py (1347→1505)
over MAX_FIRST_PARTY_FILE_LINES, failing
test_first_party_code_files_stay_under_line_limit on shard 8.

Extract the monitor-action vocabulary (AbortReason, the action dataclasses,
MonitorAction, BOT_REVIEWER_LOGINS) into a new pr_monitor_actions sibling and
re-export it from pr_monitor — mirroring the existing pr_monitor_models split —
so call sites and `import *` consumers are unchanged. Move the Docker
daemon-error decide() tests into a new test_pr_monitor_part_007 module.
